debian: HACK: overcome the lack of -lm at upstream
This patch fixes next INSTEAD bug in Debian:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=713680

Error is next:
/usr/bin/ld: SDL_rotozoom.o: undefined reference to symbol 'floor@@GLIBC_2.2.5'
/lib/x86_64-linux-gnu/libm.so.6: error adding symbols: DSO missing from command line

So we need to make sure that "SDL_rotozoom.o" file is being linked with
"libm.so" library. The only way to do that is to pass LDFLAGS variable from
"debian/rules" to "configure.sh" script. But this is hacky workaround
and should be avoided in future.

TODO: fix it in upstream.

NOTE: why LDFLAGS is not being exported at debian/rules and just added?
      See https://wiki.debian.org/HardeningWalkthrough#debhelper_9 :
      "There's no need to export CFLAGS and LDFLAGS etc. because they are
      automatically passed during calls to dh_auto_* programs"

Signed-off-by: Sam Protsenko <joe.skb7@gmail.com>